moved AddDelegation/GetTotalValWeights
